edu-match-pro-backend/alembic/env.py
from logging.config import fileConfig
import os
from dotenv import load_dotenv

from sqlalchemy import engine_from_config
from sqlalchemy import pool
from sqlmodel import SQLModel

from alembic import context

# 載入環境變數
load_dotenv()

# this is the Alembic Config object, which provides
# access to the values within the .ini file in use.
config = context.config

# Interpret the config file for Python logging.
# This line sets up loggers basically.
if config.config_file_name is not None:
    fileConfig(config.config_file_name)

# 匯入所有模型以確保它們被註冊到 SQLModel.metadata
from app.models import *

# 設定 target_metadata 為 SQLModel.metadata
target_metadata = SQLModel.metadata

# other values from the config, defined by the needs of env.py,
# can be acquired:
# my_important_option = config.get_main_option("my_important_option")
# ... etc.


def run_migrations_offline() -> None:
    """Run migrations in 'offline' mode.

    This configures the context with just a URL
    and not an Engine, though an Engine is acceptable
    here as well.  By skipping the Engine creation
    we don't even need a DBAPI to be available.

    Calls to context.execute() here emit the given string to the
    script output.

    """
    url = os.getenv("DATABASE_URL_SYNC")
    context.configure(
        url=url,
        target_metadata=target_metadata,
        literal_binds=True,
        dialect_opts={"paramstyle": "named"},
    )

    with context.begin_transaction():
        context.run_migrations()


def run_migrations_online() -> None:
    """Run migrations in 'online' mode.

    In this scenario we need to create an Engine
    and associate a connection with the context.

    """
    # 從環境變數獲取資料庫 URL，支持同步和異步格式
    database_url = os.getenv("DATABASE_URL_SYNC")
    if not database_url:
        # 如果沒有設置 DATABASE_URL_SYNC，嘗試使用 DATABASE_URL 並轉換為同步格式
        async_url = os.getenv("DATABASE_URL")
        if async_url:
            # 將 postgresql+asyncpg:// 轉換為 postgresql://
            database_url = async_url.replace("postgresql+asyncpg://", "postgresql://")
        else:
            raise ValueError("Neither DATABASE_URL_SYNC nor DATABASE_URL environment variable is set")
    
    # 設定配置
    configuration = config.get_section(config.config_ini_section, {})
    configuration["sqlalchemy.url"] = database_url
    
    connectable = engine_from_config(
        configuration,
        prefix="sqlalchemy.",
        poolclass=pool.NullPool,
    )

    with connectable.connect() as connection:
        context.configure(
            connection=connection, target_metadata=target_metadata
        )

        with context.begin_transaction():
            context.run_migrations()


if context.is_offline_mode():
    run_migrations_offline()
else:
    run_migrations_online()
